# MVP/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: >-
    Bounded machine-readable API schema for the Flask backend /api/v1 HTTP surface.
    This artifact is authoritative for route inventory and OpenAPI-tag grouping inside
    this MVP package, but it is not the sole upstream authority for all overlapping
    runtime semantics. The higher-order runtime/API authority for play-service run
    lifecycle envelopes remains docs/technical/architecture/canonical_runtime_contract.md,
    and world-engine runtime authority remains governed by docs/technical/runtime/runtime-authority-and-state-flow.md
    and docs/ADR/adr-0001-runtime-authority-in-world-engine.md.
  x-wos-governance:
    artifact_role: canonical_api_schema
    authority_scope: bounded_backend_api_v1_route_inventory
    upstream_contracts:
      - docs/technical/architecture/canonical_runtime_contract.md
      - docs/technical/runtime/runtime-authority-and-state-flow.md
      - docs/ADR/adr-0001-runtime-authority-in-world-engine.md
    not_authoritative_for:
      - world-engine fastapi surface outside documented backend projection
      - websocket message contracts
      - runtime committed truth
      - unpublished or operator-only convenience notes
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/roles:
    get:
      tags:
      - Roles
      summary: GET /roles
      operationId: get_api_v1_roles
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
    post:
      tags:
      - Roles
      summary: POST /roles
      operationId: post_api_v1_roles
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/roles/{role_id}:
    delete:
      tags:
      - Roles
      summary: DELETE /roles/{role_id}
      operationId: delete_api_v1_roles_role_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - &id022
        name: role_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
      - Roles
      summary: GET /roles/{role_id}
      operationId: get_api_v1_roles_role_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - *id022
    put:
      tags:
      - Roles
      summary: PUT /roles/{role_id}
      operationId: put_api_v1_roles_role_id
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      parameters:
      - *id022
...
  /api/v1/site/settings:
    get:
      tags:
      - SiteContent
      summary: GET /site/settings
      operationId: get_api_v1_site_settings
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
    put:
      tags:
      - SiteContent
      summary: PUT /site/settings
      operationId: put_api_v1_site_settings
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/site/slogan:
    get:
      tags:
      - SiteContent
      summary: GET /site/slogan
      operationId: get_api_v1_site_slogan
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/site/slogans:
    get:
      tags:
      - SiteContent
      summary: GET /site/slogans
      operationId: get_api_v1_site_slogans
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []